feat: api route tests for user
